requirements.txt
boto3==1.5.36
botocore==1.8.50
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
Django==2.0.2
django-cors-headers==2.1.0
django-oauth-toolkit==1.0.0
django-randomslugfield==0.3.0
django-storages==1.6.5
djangorestframework==3.7.7
docutils==0.14
idna==2.6
jmespath==0.9.3
oauthlib==2.0.6
Pillow==5.0.0
psycopg2==2.7.4
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
